Improve --log-level help text
This information was added in
7c8d4ef6b5551635daff297d3ad22e895501b0f1,
but removed again in
e27ca4e90a6c49a99e7f3d0e671d94162106d920
to make the help-output fit in a 80-chars
terminal.

This adds the available options again
in the help output, and updates the CLI
reference documentation to match actual
output.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>