Fix restartpolicy max-retry validation
the restart policy validation was moved from
the client to the daemon in 826998bed0e879df5672616acd2d7fb44ed375d2

As part of that change, retry-counts < 1
were marked as "invalid".

However, the default is 0 (unlimited), causing

    docker run -d --restart=on-failure nginx

To fail.

This changes the validation to only invalidate
retry-counts < 0.

A test was added, and other tests renamed
to allow running just these tests :)

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>