Ensure unmount before removing local volume.
When there is an error unmounting a local volume, it is still possible
to call `Remove()` on the volume causing removal of the mounted
resources which is generally not desirable.

This ensures that resources are unmounted before attempting removal.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>
(cherry picked from commit 3affa01ebe8c81101c51f71eb8ecf6cde66c3983)
Signed-off-by: Brian Goff <cpuguy83@gmail.com>